fix withdraw field balance to float

+++ cmd/gophermart/migrations/00003_create_balances_table.sql
@@ -3,7 +3,7 @@
 CREATE TABLE IF NOT EXISTS balances (
     user_id bigint not null PRIMARY KEY,
     current double precision ,
-    withdrawn integer,
+    withdrawn double precision,
     changed_at timestamptz not null DEFAULT NOW()
 );
